LPS-82165 Portlet 3.0: Move various implementation classes from package com.liferay.portlet to com.liferay.portlet.internal (CacheControlImpl)
