+++ portal-impl/src/com/liferay/portlet/MimeResponseImpl.java
@@ -19,6 +19,7 @@ import aQute.bnd.annotation.ProviderType;
 import com.liferay.portal.kernel.model.Portlet;
 import com.liferay.portal.kernel.portlet.LiferayWindowState;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portlet.internal.CacheControlImpl;
 
 import java.io.IOException;
 import java.io.OutputStream;
