+++ portal-impl/src/com/liferay/portlet/internal/CacheControlImpl.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.portlet;
+package com.liferay.portlet.internal;
 
 import com.liferay.portal.kernel.servlet.Header;
 import com.liferay.portal.kernel.servlet.MetaInfoCacheServletResponse;
+import com.liferay.portlet.MimeResponseImpl;
 
 import java.util.Collection;
 import java.util.HashMap;
